Update to less-1.3.3 Using stock javascript in this commit (which doesn't work) to show changes required from stock in next commit.
